BmiService.java changed

+++ src/BmiService.java
@@ -1,6 +1,6 @@
 public class BmiService {
     public int calculate (int mKg, double hMeter) {
-        int result = (int) ((int) mKg/(hMeter*hMeter));
+        int result = (int) (mKg/(hMeter*hMeter));
         return result;
     }
 }
